+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - new command `New-DataLayer`.
 - new command `New-DataModel`.
 - new command `New-DataFlow`.
+- new command `Select-DataJourney`.
 
 ## [0.1.0] - 2023-09-14